# Optimize BuildFQName function