@jkroepke 2025-01-21
No clue, why my contribution is even slower as the original one. The only different I saw is that I used `-count=20` as additional parameter to get and average result of multiple runs.

As I know, the performance of string concat depends on memory speed, since each string concat operation copy memory. The compiler optimization is limited on that region. 
`string.Builder` it self is described a "build string efficiently" by minimizes memory copying. That what the go team advertise. Not sure, how much overhead has `string.Builder`. It offers an byte slice with will converted to a string. string concat does similar things.

source code: 
* https://go.dev/src/runtime/string.go#L15
* https://go.dev/src/strings/builder.go



It could be possible that your memory speed on your 11th intel cpu is much higher than on my old 4th gen which result into much better results on string concats.


> Just for the record, the second benchmark of each implementation is moot, since that passes an empty name to the function, which will return very quickly due to:

If there is short-path, why the times are so different? It make no sense to me.

```
BenchmarkBuildFQNameOrig2-16            831726489                1.377 ns/op
BenchmarkBuildFQNameKroepke2-16         862387366                1.433 ns/op
BenchmarkBuildFQNameConcat2-16          1000000000               0.2183 ns/op
```